requirements.txt
requests
pydantic
colorama
ollama
chromadb
sentence-transformers

langgraph
langchain
langchain-openai
numpy
python-dotenv